🔥 Authentication pages removed (#19)
