--- .env
@@ -1,7 +1,0 @@
-API_KEY=""
-AUTH_DOMAIN=""
-PROJECT_ID=""
-STORAGE_BUCKET=""
-MESSAGING_SENDER_ID=""
-APP_ID=""
-MEASUREMENT_ID=""
