--- src/constants/routes.ts
@@ -1,5 +1,0 @@
-export const Routes = {
-  HOME: '/',
-  LOGIN: '/login',
-  SIGN_UP: '/sign-up',
-};
